+++ CHANGES.md
@@ -3,6 +3,7 @@
   - Change `assert\required` argument `$schema` as optional.
   - Add `assert\optional`.
   - Add `filter\str` functions: `nullempty` and `strip`.
+  - Add `Check` interface.
 
 ### 0.4.0 (2019-02-16)